Move test_skip_header to test/test_connection.py
It relies on sending requests that violate the HTTP spec, and that are
refused by h11 as used by Hypercorn, our new test server. Thankfully, we
don't need the return of the server, looking at what urllib3 would have
sent is enough.
